requirements.txt
# git+https://github.com/Project-MONAI/MONAI#egg=monai[nibabel, skimage, itk, tqdm, psutil, ignite, ninja]==0.8.0
monai[nibabel, skimage, itk, tqdm, psutil, ignite]==0.8.0
monailabel==0.3.0
ninja==1.10.2.3
scikit-learn==1.0.2
torchhaarfeatures==0.0.2
numpymaxflow==0.0.2